+++ apps/lite/ui/src/routes/project/$id/workspace/ResolvedOperationSource.ts
@@ -2,31 +2,15 @@ import {
 	changesInWorktreeQueryOptions,
 	commitDetailsWithLineStatsQueryOptions,
 } from "#ui/api/queries.ts";
-import {
-	commitAmendOperation,
-	commitCreateFromCommittedChangesOperation,
-	commitCreateOperation,
-	commitMoveChangesBetweenOperation,
-	commitMoveOperation,
-	commitSquashOperation,
-	commitUncommitChangesOperation,
-	commitUncommitOperation,
-	moveBranchOperation,
-	tearOffBranchOperation,
-	type Operation,
-} from "#ui/Operation.ts";
-import { createDiffSpec } from "#ui/domain/DiffSpec.ts";
 import { changeFileParent, commitFileParent, type FileParent } from "#ui/domain/FileParent.ts";
 import { QueryClient } from "@tanstack/react-query";
 import {
 	CommitDetails,
-	InsertSide,
 	WorktreeChanges,
 	type HunkHeader,
 	type TreeChange,
 } from "@gitbutler/but-sdk";
 import { Match } from "effect";
-import { decodeRefName } from "../shared";
 import { Item } from "#ui/routes/project/$id/workspace/Item.ts";
 
 type TreeChangeWithHunkHeaders = {
@@ -194,227 +178,3 @@ export const resolveOperationSource = ({
 				commitDetailsWithLineStatsQueryOptions({ projectId, commitId }).queryKey,
 			),
 	});
-
-const getCommitTargetMoveOperation = ({
-	resolvedOperationSource,
-	commitId,
-	side,
-}: {
-	resolvedOperationSource: ResolvedOperationSource;
-	commitId: string;
-	side: InsertSide;
-}) =>
-	Match.value(resolvedOperationSource).pipe(
-		Match.tags({
-			Commit: (source) =>
-				commitMoveOperation({
-					subjectCommitIds: [source.commitId],
-					relativeTo: { type: "commit", subject: commitId },
-					side,
-					dryRun: false,
-				}),
-			TreeChanges: (source) => {
-				const changes = source.changes.map(({ change, hunkHeaders }) =>
-					createDiffSpec(change, hunkHeaders),
-				);
-
-				return Match.value(source.parent).pipe(
-					Match.tagsExhaustive({
-						Change: () =>
-							commitCreateOperation({
-								relativeTo: { type: "commit", subject: commitId },
-								side,
-								changes,
-								message: "",
-								dryRun: false,
-							}),
-						Commit: (source) =>
-							commitCreateFromCommittedChangesOperation({
-								sourceCommitId: source.commitId,
-								relativeTo: { type: "commit", subject: commitId },
-								side,
-								changes,
-								dryRun: false,
-							}),
-					}),
-				);
-			},
-		}),
-		Match.orElse(() => null),
-	);
-
-const getBranchTargetOperation = ({
-	resolvedOperationSource,
-	branchRef,
-}: {
-	resolvedOperationSource: ResolvedOperationSource;
-	branchRef: Array<number>;
-}): Operation | null =>
-	Match.value(resolvedOperationSource).pipe(
-		Match.tags({
-			Branch: (source) =>
-				moveBranchOperation({
-					subjectBranch: decodeRefName(source.branchRef),
-					targetBranch: decodeRefName(branchRef),
-					dryRun: false,
-				}),
-			Commit: ({ commitId }) =>
-				commitMoveOperation({
-					subjectCommitIds: [commitId],
-					relativeTo: {
-						type: "referenceBytes",
-						subject: branchRef,
-					},
-					side: "below",
-					dryRun: false,
-				}),
-			TreeChanges: (source) => {
-				const changes = source.changes.map(({ change, hunkHeaders }) =>
-					createDiffSpec(change, hunkHeaders),
-				);
-
-				return Match.value(source.parent).pipe(
-					Match.tagsExhaustive({
-						Change: () =>
-							commitCreateOperation({
-								relativeTo: { type: "referenceBytes", subject: branchRef },
-								side: "below",
-								changes,
-								message: "",
-								dryRun: false,
-							}),
-						Commit: (source) =>
-							commitCreateFromCommittedChangesOperation({
-								sourceCommitId: source.commitId,
-								relativeTo: { type: "referenceBytes", subject: branchRef },
-								side: "below",
-								changes,
-								dryRun: false,
-							}),
-					}),
-				);
-			},
-		}),
-		Match.orElse(() => null),
-	);
-
-const getTearOffBranchTargetOperation = (
-	resolvedOperationSource: ResolvedOperationSource,
-): Operation | null => {
-	if (resolvedOperationSource._tag !== "Branch") return null;
-
-	return tearOffBranchOperation({
-		subjectBranch: decodeRefName(resolvedOperationSource.branchRef),
-		dryRun: false,
-	});
-};
-
-/**
- * | SOURCE ↓ / TARGET →    | Changes  | Commit |
- * | ---------------------- | -------- | ------ |
- * | File/hunk from changes | No-op    | Amend  |
- * | File/hunk from commit  | Uncommit | Amend  |
- * | Commit                 | Uncommit | Squash |
- *
- * Note this is currently different from the CLI's definition of "rubbing",
- * which also includes move operations.
- * https://linear.app/gitbutler/issue/GB-1160/what-should-rubbing-a-branch-into-another-branch-do#comment-db2abdb7
- */
-export const rubOperationSourceToOperation = ({
-	resolvedOperationSource,
-	target,
-}: {
-	resolvedOperationSource: ResolvedOperationSource;
-	target: Item;
-}): Operation | null =>
-	Match.value({ resolvedOperationSource, target }).pipe(
-		Match.when(
-			{ resolvedOperationSource: { _tag: "Commit" }, target: { _tag: "Commit" } },
-			({ resolvedOperationSource, target }) =>
-				commitSquashOperation({
-					sourceCommitId: resolvedOperationSource.commitId,
-					destinationCommitId: target.commitId,
-					dryRun: false,
-				}),
-		),
-		Match.when(
-			{ resolvedOperationSource: { _tag: "Commit" }, target: { _tag: "ChangesSection" } },
-			({ resolvedOperationSource }) =>
-				commitUncommitOperation({
-					commitId: resolvedOperationSource.commitId,
-					assignTo: null,
-				}),
-		),
-		Match.when(
-			{
-				resolvedOperationSource: { _tag: "TreeChanges", parent: { _tag: "Change" } },
-				target: { _tag: "Commit" },
-			},
-			({ resolvedOperationSource, target }) =>
-				commitAmendOperation({
-					commitId: target.commitId,
-					changes: resolvedOperationSource.changes.map(({ change, hunkHeaders }) =>
-						createDiffSpec(change, hunkHeaders),
-					),
-					dryRun: false,
-				}),
-		),
-		Match.when(
-			{
-				resolvedOperationSource: { _tag: "TreeChanges", parent: { _tag: "Commit" } },
-				target: { _tag: "ChangesSection" },
-			},
-			({ resolvedOperationSource }) =>
-				commitUncommitChangesOperation({
-					commitId: resolvedOperationSource.parent.commitId,
-					assignTo: null,
-					changes: resolvedOperationSource.changes.map(({ change, hunkHeaders }) =>
-						createDiffSpec(change, hunkHeaders),
-					),
-					dryRun: false,
-				}),
-		),
-		Match.when(
-			{
-				resolvedOperationSource: { _tag: "TreeChanges", parent: { _tag: "Commit" } },
-				target: { _tag: "Commit" },
-			},
-			({ resolvedOperationSource, target }) =>
-				commitMoveChangesBetweenOperation({
-					sourceCommitId: resolvedOperationSource.parent.commitId,
-					destinationCommitId: target.commitId,
-					changes: resolvedOperationSource.changes.map(({ change, hunkHeaders }) =>
-						createDiffSpec(change, hunkHeaders),
-					),
-					dryRun: false,
-				}),
-		),
-		Match.orElse(() => null),
-	);
-
-export const moveOperationSourceToOperation = ({
-	resolvedOperationSource,
-	target,
-	side,
-}: {
-	resolvedOperationSource: ResolvedOperationSource;
-	target: Item;
-	side: InsertSide;
-}) =>
-	Match.value(target).pipe(
-		Match.tags({
-			Branch: ({ branchRef }) =>
-				getBranchTargetOperation({
-					resolvedOperationSource,
-					branchRef,
-				}),
-			Commit: (target) =>
-				getCommitTargetMoveOperation({
-					resolvedOperationSource,
-					commitId: target.commitId,
-					side,
-				}),
-			BaseCommit: () => getTearOffBranchTargetOperation(resolvedOperationSource),
-		}),
-		Match.orElse(() => null),
-	);
